making form fields widgets
